简单重构netty部分代码
